C - Bit manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int num = n ^ m;
+	unsigned long int num = n ^ m;
 	int cnt = 0;
 
 	while (num > 0)
